requirements.txt
# Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Chunking
langchain-text-splitters>=0.0.1

# Embeddings
sentence-transformers>=2.2.0

# Vector store
chromadb>=0.4.0

# LLM generation
openai>=1.0.0
huggingface-hub>=0.20.0
anthropic>=0.40.0

# Reranking
transformers>=4.36.0

# UI
streamlit>=1.30.0

# Evaluation
numpy>=1.24.0
scikit-learn>=1.3.0

# Dev / testing
pytest>=7.4.0
ruff>=0.1.0
python-dotenv>=1.0.0
